refactor: movendo páginas de proposta para diferentes features
