+++ frontend/src/features/freelancers/routes/freelancer.routes.tsx
@@ -4,9 +4,6 @@ import { CadastrarProjetoFreelancer } from "@/features/freelancers/pages/Cadastr
 import { CadastrarContaBancaria } from "@/features/freelancers/pages/CadastrarContaBancaria";
 
 import { freelancerPaths } from "@/features/freelancers/routes/freelancerPaths";
-import VerProposta from "../pages/proposal/VerProposta";
-import CadastrarProposta from "../pages/proposal/CadastrarProposta";
-import MinhaProposta from "../pages/proposal/MinhaProposta";
 import { RequireRole } from "@/features/auth/components/RequireRole";
 
 export const freelancerRoutes = [
@@ -30,23 +27,6 @@ export const freelancerRoutes = [
     element: (
       <RequireRole roles={["Freelancer"]}>
         <CadastrarContaBancaria />
-      </RequireRole>
-    ),
-  },
-  { path: `${freelancerPaths.verProposta}/:id`, element: <VerProposta /> },
-  {
-    path: freelancerPaths.cadastrarProposta,
-    element: (
-      <RequireRole roles={["Freelancer"]}>
-        <CadastrarProposta />
-      </RequireRole>
-    ),
-  },
-  {
-    path: `${freelancerPaths.verMinhaProposta}/:id`,
-    element: (
-      <RequireRole roles={["Freelancer"]}>
-        <MinhaProposta />
       </RequireRole>
     ),
   },
