+++ frontend/src/features/services/dashboard/components/MiniProposalCard.tsx
