+++ frontend/src/features/services/dashboard/components/ServiceDetailsCard.tsx
