+++ frontend/src/features/services/dashboard/routes/dashboardPaths.ts
@@ -1,5 +1,7 @@
 export const dashboardServicoPaths = {
   meuServico: "/meu-servico",
+  verMinhaProposta: "/ver-minha-proposta",
 
   visualizarMeuServicoById: (id: string | number) => `/meu-servico/${id}`,
+  verMinhaPropostaById: (id: string | number) => `/ver-minha-proposta/${id}`,
 };
