+++ frontend/src/features/services/proposal/components/ProposalSection.tsx
