+++ frontend/src/features/services/proposal/components/RemovableListItem.tsx
