+++ frontend/src/features/services/proposal/components/SelectableProjectCard.tsx
