+++ frontend/src/features/services/proposal/dtos/PropostaDTO.ts
