+++ frontend/src/features/services/proposal/pages/VerProposta.tsx
@@ -9,14 +9,14 @@ import {
   consultarServicoPorId,
   aceitarProposta,
 } from "@/features/services/proposal/api/servico.api";
-import { buscarPropostaPorId } from "@/features/freelancers/api/proposta.api";
+import { buscarPropostaPorId } from "@/features/services/proposal/api/proposta.api";
 
-import type { PropostaDTO } from "@/features/freelancers/dtos/freelancer/PropostaDTO";
+import type { PropostaDTO } from "@/features/services/proposal/dtos/PropostaDTO";
 import type { FreelancerDTO } from "@/features/freelancers/dtos/freelancer/FreelancerDTO";
 import type { ServicoDTO } from "@/features/services/proposal/dtos/ServicoDTO";
 
 import { useAuth } from "@/features/auth/hooks/useAuth";
-import ProposalCards from "../../components/ProposalCards";
+import ProposalCards from "../components/ProposalCards";
 import Spinner from "@/shared/components/ui/Spinner";
 import { toLocaleString } from "@/shared/utils/date.utils";
 import { numberToCurrency } from "@/shared/utils/number.utils";
