+++ frontend/src/features/services/proposal/routes/proposal.routes.tsx
@@ -3,6 +3,8 @@ import { CadastrarServico } from "@/features/services/proposal/pages/CadastrarSe
 import PesquisaServico from "@/features/services/proposal/pages/PesquisaServicos";
 import { VisualizarServico } from "@/features/services/proposal/pages/VisualizarServico";
 import { proposalPaths } from "./proposalPaths";
+import CadastrarProposta from "../pages/CadastrarProposta";
+import VerProposta from "../pages/VerProposta";
 
 export const proposalRoutes = [
   { path: proposalPaths.pesquisaServico, element: <PesquisaServico /> },
@@ -17,6 +19,15 @@ export const proposalRoutes = [
     element: (
       <RequireRole roles={["Cliente"]}>
         <CadastrarServico />
+      </RequireRole>
+    ),
+  },
+  { path: `${proposalPaths.verProposta}/:id`, element: <VerProposta /> },
+  {
+    path: proposalPaths.cadastrarProposta,
+    element: (
+      <RequireRole roles={["Freelancer"]}>
+        <CadastrarProposta />
       </RequireRole>
     ),
   },
