+++ frontend/src/features/services/proposal/routes/proposalPaths.ts
@@ -4,4 +4,11 @@ export const proposalPaths = {
   cadastrarServico: "/cadastrar-servico",
 
   visualizarServicoById: (id: string | number) => `/visualizar-servico/${id}`,
+  
+  cadastrarProposta: "/servico/:serviceId/proposta/nova",
+  verProposta: "/ver-proposta",
+
+  cadastrarPropostaById: (servicoId: string | number) =>
+    `/servico/${servicoId}/proposta/nova`,
+  verPropostaById: (id: string | number) => `/ver-proposta/${id}`,
 };
